Python script to count rows in your exported CSV
This simple Python script checks if all items in your ZOTERO library were exported to the CSV file. To run the script, put the exported CSV file in the same directory and name it "TestLibrary.csv". Make sure that you have installed Python 3. Then open the script in the IDLE editor for Python (or similar) and run it. The script will retrieve the number of items in your library + 1 (as top row in CSV files lists documents keys).